requirements.txt
async-timeout==5.0.1
certifi==2024.12.14
charset-normalizer==3.4.0
feedgen==1.0.0
idna==3.10
lxml==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
spotipy==2.24.0
urllib3==2.2.3
